+++ NEWS.md
@@ -7,6 +7,7 @@
 * Fix bug in AllUpstreamSubids() that resulted in incorrect upstream SUBIDs when MAINPART == 0 in BranchData.txt
 * Add plots for modeled organic carbon in PlotBasinOutput() and PlotBasinSummary()
 * Update ReadGeoData() to check and warn for duplicate column names
+* Add check to PlotSubbasinRouting() to ensure that "map" and "gd" have shared SUBIDs and add handling for if SUBIDs in "map" and "gd" are not identical
 
 ## HYPEtools 1.6.4 (2024-11-27)
 *Enhancements and bug fixes*